Alert

+++ components/FavoriteComponent.js
@@ -1,6 +1,6 @@
   
 import React, { Component } from 'react';
-import { FlatList, View, Text } from 'react-native';
+import { FlatList, View, Text, Alert } from 'react-native';
 import { ListItem } from 'react-native-elements';
 import { connect } from 'react-redux';
 import { Loading } from './LoadingComponent';
@@ -33,7 +33,25 @@ class Favorites extends Component {
                 {
                     text: 'Delete', 
                     type: 'delete',
-                    onPress: () => this.props.deleteFavorite(item.id)
+                    onPress: () => {
+                        Alert.alert(
+                            'Delete Favorite?',
+                            `Are you sure you want to delete ${item.name} from favorites?`,
+                            [
+                                {
+                                    text : 'Cancel',
+                                    onPress: () => console.log("Not deleted"),
+                                    style: 'cancel'
+                                },
+                                {
+                                    text: 'OK',
+                                    onPress: ()=> this.props.deleteFavorite(item.id)
+                                }
+                            ],
+                            { cancelable: true}
+                        )
+                    }
+                    
                 }
             ];
